64bit sys support - 1185713
git-svn-id: https://nagiosplug.svn.sourceforge.net/svnroot/nagiosplug/nagiosplug/trunk@1183 f882894a-f735-0410-b71e-b25c423dba1c

+++ plugins/check_time.c
@@ -33,7 +33,7 @@ enum {
 
 #define	UNIX_EPOCH 2208988800UL
 
-unsigned long server_time, raw_server_time;
+uint32_t server_time, raw_server_time;
 time_t diff_time;
 int warning_time = 0;
 int check_warning_time = FALSE;
